factor into io and policy
